+++ CHANGELOG.md
@@ -26,6 +26,7 @@ Note that since we don't clearly distinguish between a public and private interf
   - Add easing support in camera transtion
 - Non-covalent interactions: water bridge support
 - Download Structure From AlphaFoldDB allows IDs with version suffix (version is ignored)
+- Add `loadUrl` method and GET params to Viewer app
 - Add `extensions/plugin` with several QoL improvements
   - Standalone useful loading functions previously avaiable only in the `Viewer` class
   - Standalong plugin interactivity helper function previously available only via the `Viewer` class